Update Студентам-и-поступающим.md
The link to an article is changed, so that readers can not only learn about the challenges of looking for jobs after graduation, but also about possible solutions. The new article also provides a more up to date overview of the situation (2024 instead of the old link to a 2019 article).